Remove legacy code and files related to old Go versions.
