--- request_context_1_6.go
@@ -1,14 +1,0 @@
-//go:build !go1.7
-// +build !go1.7
-
-package pubnub
-
-import (
-	"net/http"
-)
-
-func setRequestContext(r *http.Request, ctx Context) *http.Request {
-	r.Cancel = ctx.Done()
-
-	return r
-}
