+++ subscribe_context_helper.go
@@ -1,6 +1,3 @@
-//go:build go1.7
-// +build go1.7
-
 package pubnub
 
 import (
