--- subscribe_context_helper_1_6.go
@@ -1,16 +1,0 @@
-//go:build !go1.7
-// +build !go1.7
-
-package pubnub
-
-import "context"
-
-// import (
-// 	"golang.org/x/net/context"
-// )
-
-func contextWithCancel(parent context.Context) (context.Context, context.CancelFunc) {
-	return context.WithCancel(parent)
-}
-
-var backgroundContext = context.Background()
