--- utils/string_utils_1_5.go
@@ -1,4 +1,0 @@
-//go:build go1.5
-// +build go1.5
-
-package utils
