--- utils/string_utils_1_7.go
@@ -1,4 +1,0 @@
-//go:build go1.7
-// +build go1.7
-
-package utils
